Organize file structure a bit.
Start setup for users providing their own ReusableViews.
Remove setting the text on an Axis View on apply layout since this happens on every bounds change (so a ton while scrolling).
