+++ CollectionGraph/Attributes/BarGraphLayoutAttributesModel.swift
