+++ CollectionGraph/ReusableViews/DefaultLabelReusableView.swift
@@ -1,0 +1,29 @@
+
+import UIKit
+
+open class DefaultLabelReusableView: AxisReusableView {
+    
+    public let label: UILabel = UILabel()
+    
+    public override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        label.font = UIFont.systemFont(ofSize: 12)
+        label.textAlignment = .center
+        label.translatesAutoresizingMaskIntoConstraints = false
+        addSubview(label)
+        addConstraintsToLabel()
+    }
+    
+    func addConstraintsToLabel() {
+        label.centerXAnchor.constraint(equalTo: centerXAnchor).isActive = true
+        label.widthAnchor.constraint(equalTo: widthAnchor, multiplier: 1).isActive = true
+        label.centerYAnchor.constraint(equalTo: centerYAnchor).isActive = true
+        label.heightAnchor.constraint(equalTo: heightAnchor, multiplier: 1).isActive = true
+    }
+    
+    required public init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+}
